+++ src/include/sll/api/_generated.h
@@ -763,6 +763,35 @@
 
 
 
+#define __SLL_API_TYPE_sll_api_process_get_pid __SLL_CHECK_OUTPUT sll_integer_t
+#define __SLL_API_ARGS_sll_api_process_get_pid void
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_process_get_pid
+ * \group process-api
+ * \desc Docs!
+ * \ret sll_integer_t
+ */
+/**
+ * \flags check_output func optimizable
+ * \name sll_api_process_get_pid_raw
+ * \group raw-api
+ * \subgroup raw-api-process-api
+ * \desc Wrapper function for :sll_api_process_get_pid:
+ * \arg sll_object_t*const* al -> Arguments
+ * \arg sll_arg_count_t all -> Argument count
+ * \ret sll_object_t* -> The return value of the function
+ */
+/**
+ * \flags subgroup
+ * \name Process
+ * \group raw-api
+ * \subgroup raw-api-process-api
+ * \desc Docs!
+ */
+
+
+
 #define __SLL_API_TYPE_sll_api_sort_sort void
 #define __SLL_API_ARGS_sll_api_sort_sort sll_array_t* a,sll_integer_t b,sll_array_t* out
 /**