+++ src/include/sll/api/process.h
@@ -1,0 +1,26 @@
+#ifndef __SLL_API_PROCESS_H__
+#define __SLL_API_PROCESS_H__ 1
+#include <sll/api.h>
+#include <sll/common.h>
+#include <sll/types.h>
+/**
+ * \flags group
+ * \name Process API
+ * \group process-api
+ * \desc Docs!
+ */
+
+
+
+/**
+ * \flags api func optimizable
+ * \name sll_api_process_get_pid
+ * \group process-api
+ * \desc Docs!
+ * \ret I
+ */
+__API_FUNC_DECL(process_get_pid);
+
+
+
+#endif