+++ src/include/sll/platform.h
@@ -24,9 +24,23 @@
  */
 /**
  * \flags subgroup
+ * \name External Libraries
+ * \group platform
+ * \subgroup platform-lib
+ * \desc Docs!
+ */
+/**
+ * \flags subgroup
  * \name Paths
  * \group platform
  * \subgroup platform-path
+ * \desc Docs!
+ */
+/**
+ * \flags subgroup
+ * \name Processes
+ * \group platform
+ * \subgroup platform-process
  * \desc Docs!
  */
 
@@ -242,6 +256,18 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_string_length_t sll_platform_get_executabl
 
 /**
  * \flags check_output func
+ * \name sll_platform_get_pid
+ * \group platform
+ * \subgroup platform-process
+ * \desc Docs!
+ * \ret sll_pid_t
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_pid_t sll_platform_get_pid(void);
+
+
+
+/**
+ * \flags check_output func
  * \name sll_platform_list_directory
  * \group platform
  * \subgroup platform-path
@@ -272,6 +298,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_array_length_t sll_platform_list_directory
  * \flags check_output func
  * \name sll_platform_load_library
  * \group platform
+ * \subgroup platform-lib
  * \desc Docs!
  * \arg const sll_char_t* fp
  * \ret sll_library_handle_t
@@ -284,6 +311,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_library_handle_t sll_platform_load_library
  * \flags check_output func
  * \name sll_platform_lookup_function
  * \group platform
+ * \subgroup platform-lib
  * \desc Docs!
  * \arg sll_library_handle_t h
  * \arg const sll_char_t* nm
@@ -369,6 +397,7 @@ __SLL_EXTERNAL void sll_platform_sleep(sll_time_t tm);
  * \flags func
  * \name sll_platform_unload_library
  * \group platform
+ * \subgroup platform-lib
  * \desc Docs!
  * \arg sll_library_handle_t h
  */