+++ src/sll/platform/windows/win_process.c
@@ -1,0 +1,9 @@
+#include <windows.h>
+#include <sll/common.h>
+#include <sll/types.h>
+
+
+
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_pid_t sll_platform_get_pid(void){
+	return GetCurrentProcessId();
+}